:tructor: rename engine_factory
